+++ eas-20210701/src/Models/DeleteResourceInstanceLabelRequest.php
@@ -1,0 +1,104 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Eas\V20210701\Models;
+
+use AlibabaCloud\Dara\Model;
+
+class DeleteResourceInstanceLabelRequest extends Model
+{
+    /**
+     * @var bool
+     */
+    public $allInstances;
+    /**
+     * @var string[]
+     */
+    public $instanceIds;
+    /**
+     * @var string[]
+     */
+    public $keys;
+    protected $_name = [
+        'allInstances' => 'AllInstances',
+        'instanceIds'  => 'InstanceIds',
+        'keys'         => 'Keys',
+    ];
+
+    public function validate()
+    {
+        if (\is_array($this->instanceIds)) {
+            Model::validateArray($this->instanceIds);
+        }
+        if (\is_array($this->keys)) {
+            Model::validateArray($this->keys);
+        }
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->allInstances) {
+            $res['AllInstances'] = $this->allInstances;
+        }
+
+        if (null !== $this->instanceIds) {
+            if (\is_array($this->instanceIds)) {
+                $res['InstanceIds'] = [];
+                $n1                 = 0;
+                foreach ($this->instanceIds as $item1) {
+                    $res['InstanceIds'][$n1++] = $item1;
+                }
+            }
+        }
+
+        if (null !== $this->keys) {
+            if (\is_array($this->keys)) {
+                $res['Keys'] = [];
+                $n1          = 0;
+                foreach ($this->keys as $item1) {
+                    $res['Keys'][$n1++] = $item1;
+                }
+            }
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['AllInstances'])) {
+            $model->allInstances = $map['AllInstances'];
+        }
+
+        if (isset($map['InstanceIds'])) {
+            if (!empty($map['InstanceIds'])) {
+                $model->instanceIds = [];
+                $n1                 = 0;
+                foreach ($map['InstanceIds'] as $item1) {
+                    $model->instanceIds[$n1++] = $item1;
+                }
+            }
+        }
+
+        if (isset($map['Keys'])) {
+            if (!empty($map['Keys'])) {
+                $model->keys = [];
+                $n1          = 0;
+                foreach ($map['Keys'] as $item1) {
+                    $model->keys[$n1++] = $item1;
+                }
+            }
+        }
+
+        return $model;
+    }
+}