+++ eas-20210701/src/Models/UpdateResourceInstanceLabelRequest.php
@@ -1,0 +1,102 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Eas\V20210701\Models;
+
+use AlibabaCloud\Dara\Model;
+
+class UpdateResourceInstanceLabelRequest extends Model
+{
+    /**
+     * @var bool
+     */
+    public $allInstances;
+    /**
+     * @var string[]
+     */
+    public $instanceIds;
+    /**
+     * @var string[]
+     */
+    public $labels;
+    protected $_name = [
+        'allInstances' => 'AllInstances',
+        'instanceIds'  => 'InstanceIds',
+        'labels'       => 'Labels',
+    ];
+
+    public function validate()
+    {
+        if (\is_array($this->instanceIds)) {
+            Model::validateArray($this->instanceIds);
+        }
+        if (\is_array($this->labels)) {
+            Model::validateArray($this->labels);
+        }
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->allInstances) {
+            $res['AllInstances'] = $this->allInstances;
+        }
+
+        if (null !== $this->instanceIds) {
+            if (\is_array($this->instanceIds)) {
+                $res['InstanceIds'] = [];
+                $n1                 = 0;
+                foreach ($this->instanceIds as $item1) {
+                    $res['InstanceIds'][$n1++] = $item1;
+                }
+            }
+        }
+
+        if (null !== $this->labels) {
+            if (\is_array($this->labels)) {
+                $res['Labels'] = [];
+                foreach ($this->labels as $key1 => $value1) {
+                    $res['Labels'][$key1] = $value1;
+                }
+            }
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['AllInstances'])) {
+            $model->allInstances = $map['AllInstances'];
+        }
+
+        if (isset($map['InstanceIds'])) {
+            if (!empty($map['InstanceIds'])) {
+                $model->instanceIds = [];
+                $n1                 = 0;
+                foreach ($map['InstanceIds'] as $item1) {
+                    $model->instanceIds[$n1++] = $item1;
+                }
+            }
+        }
+
+        if (isset($map['Labels'])) {
+            if (!empty($map['Labels'])) {
+                $model->labels = [];
+                foreach ($map['Labels'] as $key1 => $value1) {
+                    $model->labels[$key1] = $value1;
+                }
+            }
+        }
+
+        return $model;
+    }
+}