+++ eas-20210701/ChangeLog.md
@@ -1,3 +1,12 @@
+2025-02-05 Version: 6.5.0
+- Support API DeleteResourceInstanceLabel.
+- Support API UpdateResourceInstanceLabel.
+- Update API CreateResource: update param body.
+- Update API CreateResourceInstances: update param body.
+- Update API ListResourceInstances: add param Label.
+- Update API ListResourceInstances: update response param.
+
+
 2025-01-21 Version: 6.4.2
 - Update API CreateVirtualResource: update param body.
 - Update API DescribeVirtualResource: update response param.